♻️ #235 refactor: 보스톡 댓글 작성 Response 값 변경 및 로직 수정
