+++ src/main/java/kr/co/teacherforboss/converter/CommentConverter.java
@@ -5,24 +5,21 @@ import kr.co.teacherforboss.domain.CommentLike;
 import kr.co.teacherforboss.domain.Member;
 import kr.co.teacherforboss.domain.Post;
 import kr.co.teacherforboss.domain.enums.BooleanType;
-import kr.co.teacherforboss.web.dto.CommentRequestDTO;
-import kr.co.teacherforboss.web.dto.CommentResponseDTO;
+import kr.co.teacherforboss.web.dto.BoardRequestDTO;
+import kr.co.teacherforboss.web.dto.BoardResponseDTO;
 
 import java.time.LocalDateTime;
 
 public class CommentConverter {
 
-    public static CommentResponseDTO.SaveCommentResultDTO toSaveCommentResultDTO(Comment comment) {
-        Long parentId = comment.getParent() != null ? comment.getParent().getId() : null;
-
-        return CommentResponseDTO.SaveCommentResultDTO.builder()
-                .postId(comment.getPost().getId())
-                .parentId(parentId)
+    public static BoardResponseDTO.SaveCommentDTO toSaveCommentDTO(Comment comment) {
+        return BoardResponseDTO.SaveCommentDTO.builder()
+                .commentId(comment.getId())
                 .createdAt(comment.getCreatedAt())
                 .build();
     }
 
-    public static Comment toCommentDTO(CommentRequestDTO.SaveCommentDTO request, Member member, Post post, Comment comment) {
+    public static Comment toCommentDTO(BoardRequestDTO.SaveCommentDTO request, Member member, Post post, Comment comment) {
         return Comment.builder()
                 .post(post)
                 .member(member)
@@ -33,9 +30,9 @@ public class CommentConverter {
                 .build();
     }
 
-    public static CommentResponseDTO.SaveCommentLikedResultDTO toSaveCommentLikedResultDTO(CommentLike liked) {
+    public static BoardResponseDTO.SaveCommentLikedDTO toSaveCommentLikedDTO(CommentLike liked) {
         Boolean likeStatus = (liked.getLiked() != null) ? liked.getLiked().isIdentifier() : null;
-        return CommentResponseDTO.SaveCommentLikedResultDTO.builder()
+        return BoardResponseDTO.SaveCommentLikedDTO.builder()
                 .liked(likeStatus)
                 .updatedAt(LocalDateTime.now())
                 .build();
