+++ src/main/java/kr/co/teacherforboss/service/commentService/CommentCommandService.java
@@ -1,8 +1,8 @@
 package kr.co.teacherforboss.service.commentService;
 
 import kr.co.teacherforboss.domain.Comment;
-import kr.co.teacherforboss.web.dto.CommentRequestDTO;
+import kr.co.teacherforboss.web.dto.BoardRequestDTO;
 
 public interface CommentCommandService {
-    Comment saveComment(Long postId, CommentRequestDTO.SaveCommentDTO request);
+    Comment saveComment(Long postId, BoardRequestDTO.SaveCommentDTO request);
 }
