--- src/main/java/kr/co/teacherforboss/web/dto/CommentResponseDTO.java
@@ -1,30 +1,0 @@
-package kr.co.teacherforboss.web.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-public class CommentResponseDTO {
-
-    @Getter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class SaveCommentResultDTO {
-        Long postId;
-        Long parentId;
-        LocalDateTime createdAt;
-    }
-
-    @Getter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class SaveCommentLikedResultDTO {
-        Boolean liked;
-        LocalDateTime updatedAt;
-    }
-}
